# microservices/time-series-analytics/docs/user-guide/_assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Time Series Analytics Microservice API",
    "version": "2026.2.0"
  },
  "servers": [
    {
      "url": "/ts-api"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Get the health status of the Kapacitor daemon.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    },
    "/opcua_alerts": {
      "post": {
        "summary": "Receive Alert",
        "description": "Receive and process OPC UA alerts.\n\nThis endpoint accepts alert messages in JSON format and forwards them to the \nconfigured OPC UA client.\nIf the OPC UA client is not initialized, it will attempt to initialize it \nusing the current configuration.\n\nRequest Body Example:\n    {\n        \"alert\": \"message\"\n    }\n\nResponses:\n    200:\n        description: Alert received and processed successfully.\n        content:\n            application/json:\n                example:\n                    {\n                        \"status_code\": 200,\n                        \"status\": \"success\",\n                        \"message\": \"Alert received\"\n                    }\n    '400':\n        description: OPC UA alerts are not configured in the service\n        content:\n            application/json:\n                example:\n                    {\n                        \"detail\": \"OPC UA alerts are not configured in the service\"\n                    }\n    500:\n        description: Failed to process the alert due to server error or misconfiguration.\n        content:\n            application/json:\n                example:\n                    {\n                        \"detail\": \"Failed to initialize OPC UA client: \u003Cerror_message\u003E\"\n                    }\n\nRaises:\n    HTTPException: If OPC UA alerts are not configured or if there is an error during \n    processing.",
        "operationId": "receive_alert_opcua_alerts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpcuaAlertsMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/input": {
      "post": {
        "summary": "Receive Data",
        "description": "Receives a data point in JSON format, converts it to InfluxDB line protocol, \nand sends it to the Kapacitor service.\n\nThe input JSON must include:\n    - topic (str): The topic name.\n    - tags (dict): Key-value pairs for tags (e.g., {\"location\": \"factory1\"}).\n    - fields (dict): Key-value pairs for fields (e.g., {\"temperature\": 23.5}).\n    - timestamp (int, optional): Epoch time in nanoseconds. If omitted, current time is used.\n\nExample request body:\n{\n    \"topic\": \"sensor_data\",\n    \"tags\": {\"location\": \"factory1\", \"device\": \"sensorA\"},\n    \"fields\": {\"temperature\": 23.5, \"humidity\": 60},\n    \"timestamp\": 1718000000000000000\n}\n\nArgs:\n    data_point (DataPoint): The data point to be processed, provided in the request body.\nReturns:\n    dict: A status message indicating success or failure.\nRaises:\n    HTTPException: If the Kapacitor service returns an error or if any exception \n    occurs during processing.\n\nresponses:\n    '200':\n    description: Data successfully sent to the Time series Analytics microservice\n    content:\n        application/json:\n        schema:\n            type: object\n            properties:\n            status:\n                type: string\n                example: success\n            message:\n                type: string\n                example: Data sent to Time series Analytics microservice\n    '503':\n        description: Kapacitor daemon is not running\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        detail:\n                            type: string\n                            example: \"Kapacitor daemon is not running\"\n    '4XX':\n    description: Client error (e.g., invalid input or Kapacitor error)\n    content:\n        application/json:\n        schema:\n            $ref: '#/components/schemas/HTTPValidationError'\n    '500':\n    description: Internal server error\n    content:\n        application/json:\n        schema:\n            type: object\n            properties:\n            detail:\n                type: string",
        "operationId": "receive_data_input_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataPoint"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/config": {
      "get": {
        "summary": "Get Config",
        "description": "Endpoint to retrieve the current configuration of the input service.\nAccepts an optional 'restart' query parameter and returns the current configuration \nin JSON format.\nIf 'restart=true' is provided, the Time Series Analytics Microservice UDF deployment \nservice will be restarted before returning the configuration.\n\n---\nparameters:\n    - in: query\n      name: restart\n      schema:\n        type: boolean\n        default: false\n      description: Restart the Time Series Analytics Microservice UDF deployment if true\nresponses:\n    200:\n        description: Current configuration retrieved successfully\n        content:\n            application/json:\n                schema:\n                    type: object\n                    additionalProperties: true\n                    example:\n                        {\n                            \"udfs\": { \"name\": \"udf_name\", \"model\": \"model_name\" },\n                            \"alerts\": {}\n                        }\n    500:\n        description: Failed to retrieve configuration\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        detail:\n                            type: string\n                            example: \"Failed to retrieve configuration\"",
        "operationId": "get_config_config_get",
        "parameters": [
          {
            "name": "restart",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restart the Time Series Analytics Microservice UDF deployment if true",
              "default": false,
              "title": "Restart"
            },
            "description": "Restart the Time Series Analytics Microservice UDF deployment if true"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Config File Change",
        "description": "Endpoint to handle configuration changes.\nThis endpoint can be used to update the configuration of the input service.\nUpdates the configuration of the input service with the provided key-value pairs.\n\n---\nrequestBody:\n    required: true\n    content:\n        application/json:\n            schema:\n                type: object\n                additionalProperties: true\n            example:\n                {\n                \"udfs\": {\n                    \"name\": \"udf_name\",\n                    \"model\": \"model_name\",\n                    \"device\": \"CPU/cpu or GPU/gpu\"},\n                \"alerts\": {\n                }\n                }\nresponses:\n    200:\n        description: Configuration updated successfully\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        status:\n                            type: string\n                            example: \"success\"\n                        message:\n                            type: string\n                            example: \"Configuration updated successfully\"\n    413:\n        description: Request payload exceeds the maximum allowed size of 5 KB\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        error:\n                            type: string\n                            example: \"Request exceeds the maximum allowed payload size of 5 KB.\"\n    422:\n        description: \u003E\n            Unprocessable request - invalid or missing fields, invalid device value,\n            or UDF deployment package files are missing from the server\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        detail:\n                            type: string\n                            example: \"UDF deployment package validation failed for \u003Cudf_name\u003E.\"\n    500:\n        description: Failed to write configuration to file\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        detail:\n                            type: string\n                            example: \"Failed to write configuration to file\"",
        "operationId": "config_file_change_config_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Config"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "413": {
            "description": "Request payload exceeds the maximum allowed size of 5 KB",
            "content": {
              "application/json": {
                "example": {
                  "error": "Request exceeds the maximum allowed payload size of 5 KB."
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable request - invalid or missing fields, invalid device value, or UDF deployment package files are missing from the server",
            "content": {
              "application/json": {
                "example": {
                  "detail": "UDF deployment package validation failed for \u003Cudf_name\u003E."
                }
              }
            }
          },
          "500": {
            "description": "Failed to write configuration to file",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Failed to write configuration to file"
                }
              }
            }
          }
        }
      }
    },
    "/udfs/package": {
      "post": {
        "summary": "Adds Udf Deployment Package",
        "description": "Adds UDF deployment package.\n\n**Request body**: multipart/form-data with a single field named `file` containing the tar archive.\n\nThe tar must have the following structure (no wrapping top-level directory):\n\n.. code-block:: text\n\n    udfs/\n        \u003Cudf_name\u003E.py          (required)\n        requirements.txt       (optional)\n    tick_scripts/\n        \u003Cudf_name\u003E.tick        (required)\n    models/                    (optional)\n        \u003Cmodel_files\u003E\n\n**Extraction destination**:\n\n- If `SAMPLE_APP` env var is set → `/tmp/\u003CSAMPLE_APP\u003E/`\n- Otherwise → `/tmp/\u003Ctar_filename_without_extension\u003E/`\n\n**Allowed file extensions**: `.py`, `.tick`, `.txt`, `.cb`, `.pkl`, `.json`\n`.joblib`, `.xml`, `.bin`, `.onnx`, `.pt`, `.pth`\n\nresponses:\n    200:\n        description: UDF deployment package uploaded and extracted successfully\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        status:\n                            type: string\n                            example: \"success\"\n                        message:\n                            type: string\n                            example: \"UDF deployment package 'my_udf.tar' uploaded successfully.\"\n    400:\n        description: \u003E\n            Invalid upload — file is not a .tar, archive is corrupt, failed security\n            scan (path traversal, symlink, disallowed extension),\n            or required folders/files are missing\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        detail:\n                            type: string\n                            example: \"Tar archive must contain a 'udfs/' folder with at least one .py file.\"\n    413:\n        description: Uploaded file exceeds the maximum allowed size\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        detail:\n                            type: string\n                            example: \"Uploaded file exceeds the maximum allowed size of 500 MB.\"\n    500:\n        description: Server failed to extract the UDF deployment package\n        content:\n            application/json:\n                schema:\n                    type: object\n                    properties:\n                        detail:\n                            type: string\n                            example: \"Failed to extract UDF deployment package.\"",
        "operationId": "udfs_package_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_adds_udf_deployment_package"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "400": {
            "description": "Invalid file — not a .tar, corrupt archive, failed security scan (path traversal, symlink, encrypted payload, tar-bomb expansion), or missing required folders",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Tar archive must contain a 'udfs/' folder with at least one .py file."
                }
              }
            }
          },
          "413": {
            "description": "Uploaded file exceeds the maximum allowed size",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Uploaded file exceeds the maximum allowed size of 100 MB."
                }
              }
            }
          },
          "500": {
            "description": "Failed to extract the UDF deployment package on the server",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Failed to extract UDF deployment package."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_adds_udf_deployment_package": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_adds_udf_deployment_package"
      },
      "Config": {
        "properties": {
          "udfs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Udfs",
            "default": {
              "name": "udf_name",
              "device": "CPU"
            }
          },
          "alerts": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alerts",
            "default": {

            }
          }
        },
        "type": "object",
        "title": "Config",
        "description": "Configuration model for the service."
      },
      "DataPoint": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "tags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Fields"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "fields"
        ],
        "title": "DataPoint",
        "description": "Data point model for input data."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OpcuaAlertsMessage": {
        "properties": {

        },
        "additionalProperties": true,
        "type": "object",
        "title": "OpcuaAlertsMessage",
        "description": "Model for OPC UA alert messages."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}